Use git commit-date for date in title
Note the clunky 'cut -c 2-' bit is to strip off the leading backslash.
I couldn't figure out a better way to deal with that, since I need to
escape the % char in the git-log command, but the escape char still gets
passed through to the command and ends up in the .info file.

Currently the date shows up like:

   Thu, 2 Feb 2017 16:02:08 -0500

maybe we want to tweak that.

Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>